added JSDoc comments to created components
